Fixed OverflowException crash on 64bit systems reported by sebybv.  Fix doesn't seem to cause errors on 32bit systems.
git-svn-id: https://xptable.svn.sourceforge.net/svnroot/xptable/trunk@116 907c5300-f82f-0410-99ae-ed938deb68d4

+++ Editors/KeyMessageFilter.cs
@@ -102,7 +102,7 @@ namespace XPTable.Editors
 			uiPermission.Demand();
 			Control target = Control.FromChildHandle(m.HWnd);
 
-			return this.Client.ProcessKeyMessage(target, (WindowMessage) m.Msg, m.WParam.ToInt32(), m.LParam.ToInt32());
+			return this.Client.ProcessKeyMessage(target, (WindowMessage) m.Msg, m.WParam.ToInt64(), m.LParam.ToInt64());
 		}
 	}
 }
